Replace Lombok with Record
Lombok was initially used to reduce boilerplate code
before Record functionality was introduced in Java.
Now that we have records, we can switch out Lombok for
Java records and still have no boilerplate code in the
Coffee class. In general, the fewer dependencies that
we have the more the learner can focus on the core
teaching.