Guide update (#16)
* Add a main method class to Initial folder

A Java class with a main method class is required to build the application
with CI. Additionally, the Initial folder is meant to mirror the
content that is donwloaded from Spring Initializr, which comes
with a main method class.

* Add CI file for GitHub Actions

* Include Docker Compose Support

This commit implements Spring Boot Docker Compose Support
to run the externaly Redis container. This simplifies the
learning experince and makes the guide more repeatable.

* Rewrite content about running the example

The new content will walk the user through 4 different
ways to build the application. First through a standalone jar file,
next through a docker container, next through a native compile,
and finally through a native container.

* Replace Lombok with Record

Lombok was initially used to reduce boilerplate code
before Record functionality was introduced in Java.
Now that we have records, we can switch out Lombok for
Java records and still have no boilerplate code in the
Coffee class. In general, the fewer dependencies that
we have the more the learner can focus on the core
teaching.

* Polish for README

* Add dependabot and remove Jenkins